+++ adapters/sigfox-vertx/src/main/java/org/eclipse/hono/adapter/sigfox/quarkus/MetricsFactory.java
@@ -23,6 +23,7 @@ import org.eclipse.hono.service.metric.MetricsTags;
 import org.eclipse.hono.util.Constants;
 
 import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.config.MeterFilter;
 import io.vertx.core.Vertx;
 
 /**
@@ -41,14 +42,18 @@ public class MetricsFactory {
         return new SigfoxProtocolAdapterProperties(adapterOptions);
     }
 
+    @Produces
+    @Singleton
+    MeterFilter commonTags() {
+        return MeterFilter.commonTags(MetricsTags.forProtocolAdapter(Constants.PROTOCOL_ADAPTER_TYPE_SIGFOX));
+    }
+
     @Singleton
     @Produces
     MicrometerBasedHttpAdapterMetrics metrics(
             final Vertx vertx,
             final MeterRegistry registry,
             final SigfoxProtocolAdapterProperties adapterProperties) {
-        // define tags before the first metric gets created in the MicrometerBasedHttpAdapterMetrics constructor
-        registry.config().commonTags(MetricsTags.forProtocolAdapter(Constants.PROTOCOL_ADAPTER_TYPE_SIGFOX));
         final var metrics = new MicrometerBasedHttpAdapterMetrics(registry, vertx);
         metrics.setProtocolAdapterProperties(adapterProperties);
         return metrics;