+++ service-base/src/main/java/org/eclipse/hono/service/quarkus/ApplicationPropertiesProducer.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2022 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+
+package org.eclipse.hono.service.quarkus;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
+import javax.inject.Singleton;
+
+import org.eclipse.hono.config.ApplicationConfigProperties;
+import org.eclipse.hono.config.quarkus.ApplicationOptions;
+
+/**
+ * A producer of {@link ApplicationConfigProperties}.
+ *
+ */
+@ApplicationScoped
+public class ApplicationPropertiesProducer {
+
+    @Produces
+    @Singleton
+    ApplicationConfigProperties applicationProperties(final ApplicationOptions appOptions) {
+        return new ApplicationConfigProperties(appOptions);
+    }
+}